requirements.txt
psycopg2-binary
hashlib
pandas